requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
aiofiles==23.2.1
python-dotenv==1.0.0
python-multipart==0.0.6
openai==1.3.0
tenacity==8.3.0 